Squashed commit of the following:
commit a400fe6730af5ed95346091149d15992f3e04a8d
Author: Weicheng <Weicheng.Hong@student.uts.edu.au>
Date:   Wed Sep 17 17:13:15 2025 +1000

    上一次提交没成功提交删除.pyc操作，这次补全

commit 618e0b85914a2f9f821caac37c34589c6026f979
Author: Weicheng <Weicheng.Hong@student.uts.edu.au>
Date:   Wed Sep 17 17:09:55 2025 +1000

    Modify:
    通过rm --cache并加入.gitignore文件
    将backend/.env和.pyc移除追踪。

commit 07272d9023902438825550c1cf8a35f549344831
Author: Weicheng <Weicheng.Hong@student.uts.edu.au>
Date:   Wed Sep 17 03:21:28 2025 +1000

    Modify: feat(ai): 接入 OpenAI 聊天接口，支持 explain/ask 功能
    - 由于openAI的key用不了我改成了自己的key测试
    - 增加了config.py文件统一读取环境变量
    - 修改了gpt_call.py和questionnaire.py，统一从config.py文件中导入已经读好的环境变量
    - 拆分部分pydantic响应模型到schemas文件夹中
    - 把default.json & questionnaires.json分开存储

commit 298f044e92d27112bdeb946c8c50a6ed7f40fdd5
Author: Weicheng <Weicheng.Hong@student.uts.edu.au>
Date:   Tue Sep 16 13:16:23 2025 +1000

    Modify: delete ./backend/dev.db to make it look clear to me

commit fc3e99a70378a598b84849121a37e6a7ddf3687e
Author: Weicheng <Weicheng.Hong@student.uts.edu.au>
Date:   Tue Sep 16 13:07:45 2025 +1000

    Modify: delete ./dev.db and ./env(which is now not in use)

commit e2e208750124ffa1a3344cc31a989a3c34b0f875
Author: Weicheng <Weicheng.Hong@student.uts.edu.au>
Date:   Tue Sep 16 12:57:47 2025 +1000

    Modify: directory structure and file changes

deleted:    backend/api/__pycache__/auth.cpython-313.pyc
 Modify:规范目录结构，使用config.py统一读取环境变量
 - 将.pyc和.env加入.gitignore并结束追踪
 - 环境变量应配置在.env中，然后从config.py统一读取，然后其他文件统一导入
