--- backend/.env
@@ -1,4 +1,0 @@
-OPENAI_API_KEY=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
-OPENAI_MODEL=gpt-4o-mini
-DATABASE_URL=sqlite:///./dev.db
-ENV=dev
