+++ backend/app/schemas/explain.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class ExplainOut(BaseModel):
+    outline: list[str]
+    explanation: str
+    checklist: list[str]
